fix(hmr): removed startServer
